requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-Blinka==8.27.0
adafruit-circuitpython-ads1x15==2.2.25
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==2.0.3
adafruit-circuitpython-typing==1.9.6
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.57.0
Adafruit-PureIO==1.1.11
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
distro==1.5.0
gpiozero==1.6.2
idna==2.10
numpy==1.19.5
paho-mqtt==1.6.1
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pyftdi==0.55.0
pyserial==3.5
python-apt==2.2.1
python-prctl==1.7
pyusb==1.2.1
requests==2.25.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.0
simplejpeg==1.6.4
six==1.16.0
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
toml==0.10.1
typing_extensions==4.9.0
urllib3==1.26.5
v4l2-python3==0.3.2
